refactor: move foe-gb-investment logical process to lib/foe-compute/process
And clean code
