+++ components/gb-forecast-cost/template.pug
@@ -5,7 +5,7 @@ div.content
         h3.title.is-3 {{$t(i18nPrefix + 'title')}}
     div.column.is-half(v-show="canPermalink")
       div.field
-        nuxt-link#permalink.level-right(v-bind:to="permaLink" exact)
+        nuxt-link#permalink.level-right(:to="permaLink" exact)
           span.icon
             i.fas.fa-link
           | &nbsp;{{$t('utils.permalink')}}
@@ -19,16 +19,16 @@ div.content
   div.columns
     div.column.is-half
       div.field
-        gb-list-select(v-bind:current="gb.key" v-on:change="changeGb($event)")
+        gb-list-select(:current="gb.key" @change="changeGb($event)")
 
   // Row 1
   div.columns
     div.column.is-half
-      b-field(v-bind:label="$t('utils.From')" label-for="previsionFrom" v-bind:type="haveError('from')")
-        b-input#previsionFrom(type="number" min="0" v-bind:max="to - 1" v-model.number="fromInput" name="previsionFrom" autocomplete="off")
+      b-field(:label="$t('utils.From')" label-for="previsionFrom" :type="haveError('from')")
+        b-input#previsionFrom(type="number" min="0" :max="to - 1" v-model.number="fromInput" name="previsionFrom" autocomplete="off")
     div.column.is-half
-      b-field(v-bind:label="$t('utils.To', { max: maxLevel })" label-for="previsionTo" v-bind:type="haveError('to')")
-        b-input#previsionTo(type="number" v-bind:min="from" v-bind:max="maxLevel" v-model.number="to" name="previsionTo" autocomplete="off")
+      b-field(:label="$t('utils.To', { max: maxLevel })" label-for="previsionTo" :type="haveError('to')")
+        b-input#previsionTo(type="number" :min="from" :max="maxLevel" v-model.number="to" name="previsionTo" autocomplete="off")
 
   // Row 2
   div.columns
@@ -36,7 +36,7 @@ div.content
       yes-no(v-model="customPercentage" :label="$t(i18nPrefix + 'custom_percentage')")
 
     div.column.is-half(v-show="!customPercentage")
-      b-field(v-bind:label="$t('components.gb_investment.gb_investment.form.investor_percentage')" label-for="previsionInvestorPercentageGlobal" v-bind:type="haveError('investorPercentageGlobal')")
+      b-field(:label="$t('components.gb_investment.gb_investment.form.investor_percentage')" label-for="previsionInvestorPercentageGlobal" :type="haveError('investorPercentageGlobal')")
         b-input#previsionInvestorPercentageGlobal(type="number" min="0" step="0.1" v-model.number="investorPercentageGlobal" name="investorPercentageGlobal" autocomplete="off")
 
   // Row 3
@@ -50,8 +50,8 @@ div.content
         tr(v-for="i in 5")
           td.has-text-right {{i}}
           td
-            b-field(v-bind:type="haveError('investorPercentageCustom_' + i)")
-              b-input(type="number" min="0" step="0.1" v-model.number="investorPercentageCustom[i - 1]" v-bind:name="'investorPercentageCustom_' + i" autocomplete="off")
+            b-field(:type="haveError('investorPercentageCustom_' + i)")
+              b-input(type="number" min="0" step="0.1" v-model.number="investorPercentageCustom[i - 1]" :name="'investorPercentageCustom_' + i" autocomplete="off")
 
   // Result
   div(v-if="previsionResult !== null")
@@ -63,14 +63,14 @@ div.content
 
   div.columns
     div.column.is-half
-      b-field(v-bind:label="$t('components.cf_calculator.block_cfc.fp_by_24h')" label-for="previsionFpBy24h" v-bind:type="haveError('fpBy24h')")
+      b-field(:label="$t('components.cf_calculator.block_cfc.fp_by_24h')" label-for="previsionFpBy24h" :type="haveError('fpBy24h')")
         b-input#previsionFpBy24h(type="number" min="0" v-model.number="fpBy24h" name="previsionFpBy24h" autocomplete="off")
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'prop_already_invested')" label-for="previsionAlreadyInvested" v-bind:type="haveError('alreadyInvested')")
-        b-input#previsionAlreadyInvested(type="number" min="0" v-bind:max="maxCurrentLevel" v-model.number="alreadyInvested" name="previsionAlreadyInvested" autocomplete="off")
+      b-field(:label="$t(i18nPrefix + 'prop_already_invested')" label-for="previsionAlreadyInvested" :type="haveError('alreadyInvested')")
+        b-input#previsionAlreadyInvested(type="number" min="0" :max="maxCurrentLevel" v-model.number="alreadyInvested" name="previsionAlreadyInvested" autocomplete="off")
 
   div(v-show="(previsionResult !== null) && (fpBy24h > 0)")
     p {{$t(i18nPrefix + 'message.estimated_time', { time: estimatedTime })}}
 
   div(v-show="from !== to")
-    graph-canvas.margin-top-one-em(v-show="from < to" gtype="line" id="graph" v-bind:goptions="options" v-bind:labels="labels" v-bind:datasets="datasets")
+    graph-canvas.margin-top-one-em(v-show="from < to" gtype="line" id="graph" :goptions="options" :labels="labels" :datasets="datasets")
