+++ components/gb-investment-investors/script.js
@@ -1,6 +1,6 @@
 import Vue from "vue";
 import Utils from "~/scripts/utils";
-import gbProcess from "~/scripts/foe-gb-investment";
+import gbProcess from "~/lib/foe-compute-process/gb-investment";
 import gbListSelect from "~/components/gb-list-select/GbListSelect";
 
 const i18nPrefix = "components.gb_investment_investors.";
