+++ components/gb-investment/script.js
@@ -1,5 +1,5 @@
 import Utils from "~/scripts/utils";
-import gbProcess from "~/scripts/foe-gb-investment";
+import gbProcess from "~/lib/foe-compute-process/gb-investment";
 import gbListSelect from "~/components/gb-list-select/GbListSelect";
 import yesNo from "~/components/yes-no/YesNo";
