+++ components/gb-investment/template.pug
@@ -4,10 +4,10 @@ div.content
       div.field
         h3.title.is-3 {{$t('foe_data.gb.' + gb.key)}}
       div.field
-        gb-list-select(v-bind:current="gb.key" v-on:change="goTo($event)")
+        gb-list-select(:current="gb.key" @change="goTo($event)")
     div.column.is-half(v-show="canPermalink")
       div.field
-        nuxt-link#permalink.level-right(v-bind:to="permaLink" exact)
+        nuxt-link#permalink.level-right(:to="permaLink" exact)
           span.icon
             i.fas.fa-link
           | &nbsp;{{$t('utils.permalink')}}
@@ -18,20 +18,20 @@ div.content
 
   div.columns
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'gb_investment.form.level', { level: levelNormalized, fromLevel: levelNormalized - 1, min: 1, max: maxLevel })" label-for="giLevel" v-bind:type="haveError('level')")
-        b-input#giLevel(type="number" min="1" v-bind:max="maxLevel" v-model.number="level" name="level" autocomplete="off" v-bind:autofocus="canPermalink")
+      b-field(:label="$t(i18nPrefix + 'gb_investment.form.level', { level: levelNormalized, fromLevel: levelNormalized - 1, min: 1, max: maxLevel })" label-for="giLevel" :type="haveError('level')")
+        b-input#giLevel(type="number" min="1" :max="maxLevel" v-model.number="level" name="level" autocomplete="off" :autofocus="canPermalink")
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'gb_investment.form.investor_percentage')" label-for="giInvestorPercentageGlobal" v-bind:type="haveError('investorPercentageGlobal')")
+      b-field(:label="$t(i18nPrefix + 'gb_investment.form.investor_percentage')" label-for="giInvestorPercentageGlobal" :type="haveError('investorPercentageGlobal')")
         b-input#giInvestorPercentageGlobal(type="number" min="0" step="0.1" v-model.number="investorPercentageGlobal" name="investorPercentageGlobal" autocomplete="off")
 
   div.columns(v-if="result !== null")
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'gb_investment.form.owner_investment', { max: result.cost })" label-for="giOwnerInvestment" v-bind:type="haveError('ownerInvestment')")
-        b-input#giOwnerInvestment(type="number" min="0" v-bind:max="result.cost" v-model.number="ownerInvestment" name="ownerInvestment" autocomplete="off")
+      b-field(:label="$t(i18nPrefix + 'gb_investment.form.owner_investment', { max: result.cost })" label-for="giOwnerInvestment" :type="haveError('ownerInvestment')")
+        b-input#giOwnerInvestment(type="number" min="0" :max="result.cost" v-model.number="ownerInvestment" name="ownerInvestment" autocomplete="off")
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'gb_investment.form.add_investors', { min: 1, max: maxInvestment })" label-for="giAddInvestors" v-bind:type="haveError('addInvestors')")
+      b-field(:label="$t(i18nPrefix + 'gb_investment.form.add_investors', { min: 1, max: maxInvestment })" label-for="giAddInvestors" :type="haveError('addInvestors')")
         b-field(grouped)
-          b-input#giAddInvestors(type="number" min="1" :max="maxInvestment" v-model.number="addInvestors" v-on:keyup.native.enter="addInvestor" name="addInvestors" autocomplete="off" expanded :disabled="maxInvestment <= 0" @click.native="haveReadTipAboutAddInvestor")
+          b-input#giAddInvestors(type="number" min="1" :max="maxInvestment" v-model.number="addInvestors" @keyup.native.enter="addInvestor" name="addInvestors" autocomplete="off" expanded :disabled="maxInvestment <= 0" @click.native="haveReadTipAboutAddInvestor")
           p.control
             button.button.is-info(:title="$t('utils.Add')" @click="addInvestor")
               span.icon
@@ -88,11 +88,11 @@ div.content
                   | )
           // Custom arc bonus
           td
-            b-field(v-bind:type="haveError('investorPercentageCustom_' + i)")
-              b-input(type="number" min="0" step="0.1" v-model.number="investorPercentageCustom[i]" v-bind:name="'investorPercentageCustom_' + i" autocomplete="off")
+            b-field(:type="haveError('investorPercentageCustom_' + i)")
+              b-input(type="number" min="0" step="0.1" v-model.number="investorPercentageCustom[i]" :name="'investorPercentageCustom_' + i" autocomplete="off")
           // Place free
           td
-            b-switch(v-bind:value="placeFree[i].state" type="is-info" v-on:input="changePlaceFree(i, $event)") {{placeFree[i].state ? $t('utils.Yes') : $t('utils.No')}}
+            b-switch(:value="placeFree[i].state" type="is-info" @input="changePlaceFree(i, $event)") {{placeFree[i].state ? $t('utils.Yes') : $t('utils.No')}}
 
         template(v-if="result.otherInvestment.length && showExtraInvestors")
           tr(v-for="(place, i) in result.otherInvestment")
@@ -167,10 +167,10 @@ div.content
                 b-tooltip(position="is-left" :type="$store.state.isDarkTheme ? 'is-light' : 'is-dark'" :label="$t(i18nPrefix + 'gb_investment.table.tbody.tooltip_expected_investment')" dashed) {{$formatNumber(props.row.participation)}}
                 | )
         b-table-column(v-if="props.index < result.investment.length" field="custom_ip" :label="$t($data.i18nPrefix + 'gb_investment.table.thead.th5')" :numeric="true")
-          b-field(v-bind:type="haveError('investorPercentageCustom_' + props.index)")
-            b-input(type="number" min="0" step="0.1" v-model.number="investorPercentageCustom[props.index]" v-bind:name="'investorPercentageCustom_' + props.index" autocomplete="off")
+          b-field(:type="haveError('investorPercentageCustom_' + props.index)")
+            b-input(type="number" min="0" step="0.1" v-model.number="investorPercentageCustom[props.index]" :name="'investorPercentageCustom_' + props.index" autocomplete="off")
         b-table-column(v-if="props.index < result.investment.length" field="place_free" :label="$t($data.i18nPrefix + 'gb_investment.table.thead.th6')")
-          b-switch(v-bind:value="placeFree[props.index].state" type="is-info" v-on:input="changePlaceFree(props.index, $event)") {{placeFree[props.index].state ? $t('utils.Yes') : $t('utils.No')}}
+          b-switch(:value="placeFree[props.index].state" type="is-info" @input="changePlaceFree(props.index, $event)") {{placeFree[props.index].state ? $t('utils.Yes') : $t('utils.No')}}
 
       template(slot="footer" v-if="result")
         th(colspan=3).has-text-centered
@@ -195,10 +195,10 @@ div.content
 
   div.columns
     div.column.is-half
-      b-field(v-bind:label="$t('utils.Prefix')" label-for="giPrefix")
+      b-field(:label="$t('utils.Prefix')" label-for="giPrefix")
         b-input#giPrefix(type="text" v-model.trim="prefix" name="prefix" autocomplete="off" :placeholder="$t($data.i18nPrefix + 'promotion.prefix_placeholder')")
     div.column.is-half
-      b-field(v-bind:label="$t('utils.Suffix')" label-for="giSuffix")
+      b-field(:label="$t('utils.Suffix')" label-for="giSuffix")
         b-input#giSuffix(type="text" v-model.trim="suffix" name="suffix" autocomplete="off" :placeholder="$t($data.i18nPrefix + 'promotion.suffix_placeholder')")
 
   div.columns
@@ -215,8 +215,8 @@ div.content
     b-field(v-for="(value, i) in promotion" :key="i")
       b-input(type="text" :value="value.message" readonly expanded)
       p.control
-        b-tooltip(:type="$store.state.isDarkTheme ? 'is-light' : 'is-dark'" v-bind:label="$t('utils.Copyed')" :active="value.active" always)
-          button.button.copy(v-clipboard="value.message" @success="successCopy(i)" v-bind:title="$t('utils.Copy')")
+        b-tooltip(:type="$store.state.isDarkTheme ? 'is-light' : 'is-dark'" :label="$t('utils.Copyed')" :active="value.active" always)
+          button.button.copy(v-clipboard="value.message" @success="successCopy(i)" :title="$t('utils.Copy')")
             span.icon
               i.far.fa-copy
             | &nbsp;&nbsp;{{$t('utils.Copy')}}
