+++ components/gb-statistics/template.pug
@@ -5,7 +5,7 @@ div.content
         h3.title.is-3 {{$t(i18nPrefix + 'title')}}
     div.column.is-half
       div.field
-        nuxt-link#permalink.level-right(v-bind:to="permaLink" exact)
+        nuxt-link#permalink.level-right(:to="permaLink" exact)
           span.icon
             i.fas.fa-link
           | &nbsp;{{$t('utils.permalink')}}
@@ -19,10 +19,10 @@ div.content
 
   div.columns
     div.column.is-half
-      b-field(v-bind:label="$t('utils.From')" label-for="gbStatisticsFrom" v-bind:type="haveError('from')")
-        b-input#gbStatisticsFrom(type="number" min="1" v-bind:max="to" v-model.number="from" name="from" autocomplete="off")
+      b-field(:label="$t('utils.From')" label-for="gbStatisticsFrom" :type="haveError('from')")
+        b-input#gbStatisticsFrom(type="number" min="1" :max="to" v-model.number="from" name="from" autocomplete="off")
     div.column.is-half
-      b-field(v-bind:label="$t('utils.To', { max: maxLevelGraph })" label-for="gbStatisticsTo" v-bind:type="haveError('to')")
-        b-input#gbStatisticsTo(type="number" v-bind:min="from" v-bind:max="maxLevelGraph" v-model.number="to" name="to" autocomplete="off")
+      b-field(:label="$t('utils.To', { max: maxLevelGraph })" label-for="gbStatisticsTo" :type="haveError('to')")
+        b-input#gbStatisticsTo(type="number" :min="from" :max="maxLevelGraph" v-model.number="to" name="to" autocomplete="off")
 
-  graph-canvas(gtype="line" id="graph" v-bind:goptions="options" v-bind:labels="labels" v-bind:datasets="datasets" @switchVisibility="switchVisibility($event)")
+  graph-canvas(gtype="line" id="graph" :goptions="options" :labels="labels" :datasets="datasets" @switchVisibility="switchVisibility($event)")
