+++ components/graph-canvas/template.pug
@@ -1,1 +1,1 @@
-canvas(v-bind:id="id" :style="$store.state.isDarkTheme ? 'background-color: white;' : ''")
+canvas(:id="id" :style="$store.state.isDarkTheme ? 'background-color: white;' : ''")
