+++ components/secure-position/template.pug
@@ -5,7 +5,7 @@ div.content
         h3.title.is-3 {{$t(i18nPrefix + 'block_place.title')}}
     div.column.is-half(v-show="canPermalink")
       div.field
-        nuxt-link#permalink.level-right(v-bind:to="permaLink" exact)
+        nuxt-link#permalink.level-right(:to="permaLink" exact)
           span.icon
             i.fas.fa-link
           | &nbsp;{{$t('utils.permalink')}}
@@ -17,29 +17,29 @@ div.content
 
   div.columns
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'block_place.level_cost')" label-for="splevelCost" v-bind:type="haveError('levelCost')")
-        b-input(type="number" min="0" v-model.number="levelCost" name="level-cost" id="splevelCost" autocomplete="off" v-bind:disabled="haveInputLevelCost()" v-bind:autofocus="canPermalink")
+      b-field(:label="$t(i18nPrefix + 'block_place.level_cost')" label-for="splevelCost" :type="haveError('levelCost')")
+        b-input(type="number" min="0" v-model.number="levelCost" name="level-cost" id="splevelCost" autocomplete="off" :disabled="haveInputLevelCost()" :autofocus="canPermalink")
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'block_place.current_deposits')" label-for="spCurrentDeposits" v-bind:type="haveError('currentDeposits')")
+      b-field(:label="$t(i18nPrefix + 'block_place.current_deposits')" label-for="spCurrentDeposits" :type="haveError('currentDeposits')")
         b-input(type="number" min="0" v-model.number="currentDeposits" name="current-deposits" id="spCurrentDeposits" autocomplete="off")
 
   div.columns
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'block_place.your_participation')" label-for="spYourParticipation" v-bind:type="haveError('yourParticipation')")
+      b-field(:label="$t(i18nPrefix + 'block_place.your_participation')" label-for="spYourParticipation" :type="haveError('yourParticipation')")
         b-input(type="number" min="0" v-model.number="yourParticipation" name="your-participation" id="spYourParticipation" autocomplete="off")
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'block_place.other_participation')" label-for="spOtherParticipation" v-bind:type="haveError('otherParticipation')")
+      b-field(:label="$t(i18nPrefix + 'block_place.other_participation')" label-for="spOtherParticipation" :type="haveError('otherParticipation')")
         b-input(type="number" min="0" v-model.number="otherParticipation" name="other-participation" id="spOtherParticipation" autocomplete="off")
 
   div.columns
     div.column.is-half
-      b-field(v-bind:label="$t(i18nPrefix + 'block_place.your_arc_bonus')" label-for="spYourArcBonus" v-bind:type="haveError('yourArcBonus')")
+      b-field(:label="$t(i18nPrefix + 'block_place.your_arc_bonus')" label-for="spYourArcBonus" :type="haveError('yourArcBonus')")
         b-input(type="number" min="0" v-model.number="yourArcBonus" step="0.1" name="your-arc-bonus" id="spYourArcBonus" autocomplete="off")
     div.column.is-half
-      b-field(v-if="haveInputLevelCost()" v-bind:label="$t(i18nPrefix + 'block_place.target_reward')" label-for="fpTargetReward" v-bind:type="haveError('fpTargetReward')")
+      b-field(v-if="haveInputLevelCost()" :label="$t(i18nPrefix + 'block_place.target_reward')" label-for="fpTargetReward" :type="haveError('fpTargetReward')")
         b-select#fpTargetReward(v-model="fpTargetReward" name="fp-target-reward" expanded)
-          option(v-for="(place, i) in levelData.investment" :value="place.reward" :key="i" v-bind:id="'option_' + i" autocomplete="off") {{$t(i18nPrefix + 'block_place.place', { place: i + 1, count: place.reward })}}
-      b-field(v-else v-bind:label="$t(i18nPrefix + 'block_place.fp_target_reward')" label-for="spFpTargetReward" v-bind:type="haveError('fpTargetReward')")
+          option(v-for="(place, i) in levelData.investment" :value="place.reward" :key="i" :id="'option_' + i" autocomplete="off") {{$t(i18nPrefix + 'block_place.place', { place: i + 1, count: place.reward })}}
+      b-field(v-else :label="$t(i18nPrefix + 'block_place.fp_target_reward')" label-for="spFpTargetReward" :type="haveError('fpTargetReward')")
         b-input(type="number" min="0" v-model.number="fpTargetReward" step="5" name="fp-target-reward" id="spFpTargetReward" autocomplete="off")
 
   p {{$t(i18nPrefix + 'block_place.fp_to_complete_level', {count: getNumberOfRemainingPoints()})}}
@@ -51,7 +51,7 @@ div.content
   p(v-else) {{$t(i18nPrefix + 'block_place.result.cant_block')}}
 
   div.margin-top-one-em(v-show="(fp > 0) && (yourArcBonus >= 0) && (fpTargetReward > 0)")
-    article.message(v-bind:class="{'is-success': roi > 0, 'is-danger': roi < 0}")
+    article.message(:class="{'is-success': roi > 0, 'is-danger': roi < 0}")
       div.message-body.has-text-dark
         p(v-if="roi > 0")
           span.icon.is-medium.has-text-success(style="vertical-align: middle;")
