+++ lib/foe-compute-process/cf-calculator.js
@@ -10,10 +10,12 @@ export default {
    * @param secondRq {boolean} True to have second quests, false otherwise
    * @param suppliesGathered {number} Your daily production of supplies
    * @param otherRq {number} Other daily accomplished quests
-   * @param cumulativeQuest {number} Number of cumulative quests. Used when we gain more coins and supplies than you use in quests.
+   * @param cumulativeQuest {number} Number of cumulative quests. Used when we gain more coins and supplies than you
+   * use in quests.
    * @returns {{bp: number, medals: number, goods: number, fp: number, dailyUbq: number,
    * bonusUbq: number, secondRqCompleted: number, totalGoods: number, totalFp: number, totalRqCompleted: number,
-   * coinSupplyReturn: Array, infinityGenerator: boolean, defaultCumulativeQuest: number}} All awards obtained statistically.
+   * coinSupplyReturn: Array, infinityGenerator: boolean, defaultCumulativeQuest: number}}
+   * All awards obtained statistically.
    */
   compute(age, yourCfBoost, coins, supplies, goods, fpBy24h, secondRq, suppliesGathered, otherRq, cumulativeQuest = 0) {
     let cf_boost = yourCfBoost / 100;
