+++ lib/foe-compute-process/gb-investment.js
@@ -1,5 +1,5 @@
-import Utils from "./utils";
-import Errors from "./errors";
+import Utils from "../../scripts/utils";
+import Errors from "../../scripts/errors";
 
 /**
  * Check if a numeric parameter are valid. Throw an error if the value are not valid.
@@ -323,6 +323,12 @@ export default {
       );
     }
 
-    return levelInvestment(gb, currentLevel, investorPercentage, defaultParticipation.sort((a, b) => b - a), ownerPreparation);
+    return levelInvestment(
+      gb,
+      currentLevel,
+      investorPercentage,
+      defaultParticipation.sort((a, b) => b - a),
+      ownerPreparation
+    );
   }
 };
