+++ pages/_lang/changelog/template.pug
@@ -1,3 +1,3 @@
 div#changelog.content(style="min-height: 100px")
   b-loading(:is-full-page="false" :active.sync="markdown.length === 0")
-  remark(v-bind:markdown="markdown")
+  remark(:markdown="markdown")
