+++ pages/_lang/gb-investment/template.pug
@@ -8,6 +8,6 @@ div
         div.content
           ul
             li(v-for="gb in item.gbs")
-              a(v-bind:href="$i18nPath('gb-investment/' + gb + '/')") {{$t('foe_data.gb.' + gb)}}
+              a(:href="$i18nPath('gb-investment/' + gb + '/')") {{$t('foe_data.gb.' + gb)}}
               // FIXME: with nuxt-link, no URL query in gb-investment
-              //nuxt-link(v-bind:to="$i18nPath('gb-investment/' + gb.key + '/')" exact) {{$t('foe_data.gb.' + gb.key)}}
+              //nuxt-link(:to="$i18nPath('gb-investment/' + gb.key + '/')" exact) {{$t('foe_data.gb.' + gb.key)}}
